requirements.txt
deepctr==0.9.0
numpy==1.15.4
pandas==0.20.3
requests==2.22.0
scikit-learn==0.20.3
tensorflow-gpu==1.12
tqdm==4.58.0